Add basic authentication mechanismes

+++ server/install.sh
@@ -108,9 +108,13 @@ install() {
     
     # Config server parameters
     websocket_server_port=""
+    websocket_server_secret=""
+    websocket_authorized_clients_ips=""
     
     # Config flags
     conf_websocket_server_port=false
+    conf_websocket_server_secret=false
+    conf_websocket_authorized_clients_ips=false
     
     # Automatic setup : try loading config file
     if [ -f "${CONFIG_FILE}" ]; then
@@ -124,7 +128,27 @@ install() {
         else
             echo "Key 'websocket_server_port' not found or has no value in ${CONFIG_FILE}"
         fi
-      
+
+        # Automatic setup of "websocket_server_secret"
+        websocket_server_secret=$(grep "^websocket_server_secret:" "${CONFIG_FILE}" | cut -d':' -f2- ) # Retrieve from file
+        websocket_server_secret=$(echo "$websocket_server_secret" | xargs) # Trims whitespace
+        if [ -n "${websocket_server_secret}" ]; then
+            conf_websocket_server_secret=true
+            echo "Using pre-configured 'websocket_server_secret' value ${websocket_server_secret} from ${CONFIG_FILE}"
+        else
+            echo "Key 'websocket_server_secret' not found or has no value in ${CONFIG_FILE}"
+        fi
+
+        # Automatic setup of "websocket_authorized_clients_ips"
+        websocket_authorized_clients_ips=$(grep "^websocket_authorized_clients_ips:" "${CONFIG_FILE}" | cut -d':' -f2- ) # Retrieve from file
+        websocket_authorized_clients_ips=$(echo "$websocket_authorized_clients_ips" | xargs) # Trims whitespace
+        if [ -n "${websocket_authorized_clients_ips}" ]; then
+            conf_websocket_authorized_clients_ips=true
+            echo "Using pre-configured 'websocket_authorized_clients_ips' value ${websocket_authorized_clients_ips} from ${CONFIG_FILE}"
+        else
+            echo "Key 'websocket_authorized_clients_ips' not found or has no value in ${CONFIG_FILE}"
+        fi
+
     else
         # Automatic setup : no config file or config file not accessible
         echo "Config file not found: ${CONFIG_FILE}"
@@ -148,18 +172,65 @@ install() {
         done
     fi
 
+    # Manual setup of "websocket_server_secret":
+    if [ "${conf_websocket_server_secret}" != "true" ]; then
+        regex='^.+$'
+        while true; do
+            read -p "Enter this USB gadget server secret (ex: myServerSecret): " websocket_server_secret </dev/tty
+
+            # trim whitespace and preserve double-quotes
+            websocket_server_secret="${websocket_server_secret#"${websocket_server_secret%%[![:space:]]*}"}"
+            websocket_server_secret="${websocket_server_secret%"${websocket_server_secret##*[![:space:]]}"}"
+
+            # Escape single-quotes
+            websocket_server_secret="${websocket_server_secret//\'/\'\\\'\'}"
+
+            if [[ "$websocket_server_secret" =~ $regex ]]; then
+                break
+            else
+                echo "Please answer a well-formed secret (non-empty and non-whitespaces-only secret expected)"
+            fi
+        done
+    fi
+
+    # Manual setup of "websocket_authorized_clients_ips":
+    if [ "${conf_websocket_authorized_clients_ips}" != "true" ]; then
+        regex='^"([0-9]{1,3}\.){3}[0-9]{1,3}"([[:space:]]*,[[:space:]]*"([0-9]{1,3}\.){3}[0-9]{1,3}")*$'
+        while true; do
+            read -p "Enter your USB gadget authorized clients IPv4 addresses (ex: \"192.168.1.15\",..,\"127.0.0.1\"): " websocket_authorized_clients_ips </dev/tty
+
+            # trim whitespace and preserve double-quotes
+            websocket_authorized_clients_ips="${websocket_authorized_clients_ips#"${websocket_authorized_clients_ips%%[![:space:]]*}"}"
+            websocket_authorized_clients_ips="${websocket_authorized_clients_ips%"${websocket_authorized_clients_ips##*[![:space:]]}"}"
+
+            if [[ ${websocket_authorized_clients_ips} =~ ${regex} ]]; then
+                break
+            else
+                echo "Please answer a well-formed list of authorized clients IPv4 addresses (ex: \"192.168.1.15\",..,\"127.0.0.1\" expected)"
+            fi
+        done
+    fi
+
     # Write updated config file
     echo "Writing config file ${CONFIG_FILE}..."
     cat <<EOF > "${CONFIG_FILE}"
 websocket_server_port: ${websocket_server_port}
+websocket_server_secret: '${websocket_server_secret}'
+websocket_authorized_clients_ips: ${websocket_authorized_clients_ips}
 EOF
 
     # Configure using new configurations
     echo "Configuring trackpad_mouse server..."
     sed -i "s|<websocket_server_port>|${websocket_server_port}|g" "${MODULE_FILE}"
     echo "This USB gadget server will be running at 0.0.0.0:${websocket_server_port}"
-    
-    
+
+    sed -i "s|<websocket_server_secret>|${websocket_server_secret}|g" "${MODULE_FILE}"
+    echo "This USB gadget server secret set to ${websocket_server_secret}"
+
+    escaped_websocket_authorized_clients_ips=$(printf '%s' "$websocket_authorized_clients_ips" | sed 's/"/\\"/g')
+    sed -i "s|<websocket_authorized_clients_ips>|${escaped_websocket_authorized_clients_ips}|g" "${MODULE_FILE}"
+    echo "This USB gadget server access authorization set to IPs ${websocket_authorized_clients_ips}"
+
     chown -R :ha_zero_hid /opt/ha_zero_hid
     chmod -R g+w /opt/ha_zero_hid
